pyproject.toml
[project]
name = "quote_bot"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "盐粒 Yanli", email = "mail@yanli.one"},
]
dependencies = ["httpx>=0.28.1", "Pillow>=11.1.0", "pilmoji>=2.0.0", "emoji>=2.10.0"]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
quote-bot = "quote_bot.bot:main"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
